@@ -1,6 +1,12 @@
-<div class="row">
-    <div class="col">
-        <h1>The Transition Pathway</h1>
-    </div>
-</div>
-<slot />
+<script>
+	/**
+	 * @typedef {Object} Props
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { children } = $props();
+</script>
+
+<h1>The Transition Pathway</h1>
+{@render children?.()}
