@@ -1,32 +1,39 @@
 <script lang="ts">
-    import { createEventDispatcher } from "svelte";
-    export let options: any[];
-    export let selected_option: any = options[0];
-    export let enabled: boolean = true;
-    let uniqueID = options.join("_");
+	interface Props {
+		options: any[];
+		selected_option?: any;
+		enabled?: boolean;
+		selection_changed: (selected_option: any) => void;
+	}
 
-    const dispatch = createEventDispatcher();
+	let {
+		options,
+		selected_option = $bindable(options[0]),
+		enabled = true,
+		selection_changed
+	}: Props = $props();
+	let uniqueID = options.join('_');
 
-    function update_selection() {
-        dispatch("selection-changed", selected_option);
-    }
+	function update_selection() {
+		selection_changed(selected_option);
+	}
 </script>
 
 <div role="radiogroup" id={`group-${uniqueID}`}>
-    {#each options as option, i}
-        <div class="form-check">
-            <input
-                class="form-check-input"
-                type="radio"
-                id={"radio-" + option}
-                bind:group={selected_option}
-                on:change={update_selection}
-                value={option}
-                disabled={!enabled}
-            />
-            <label class="form-check-label" for={"radio-" + option}>
-                {option}
-            </label>
-        </div>
-    {/each}
+	{#each options as option, i}
+		<div class="form-check">
+			<input
+				class="form-check-input"
+				type="radio"
+				id={'radio-' + option}
+				bind:group={selected_option}
+				onchange={update_selection}
+				value={option}
+				disabled={!enabled}
+			/>
+			<label class="form-check-label" for={'radio-' + option}>
+				{option}
+			</label>
+		</div>
+	{/each}
 </div>
