@@ -1,3 +1,3 @@
 <script lang="ts">
-    import { base } from "$app/paths";
+	import { base } from '$app/paths';
 </script>
