@@ -1,79 +1,76 @@
 <script lang="ts">
-	import SolutionFilter from "../../../components/SolutionFilter.svelte";
-	import type {
-		ActivatedSolution,
-		ComponentTotal,
-		Dataset,
-		Row,
-	} from "$lib/types";
-	import AllCheckbox from "../../../components/AllCheckbox.svelte";
-	import Radio from "../../../components/Radio.svelte";
-	import ToggleButton from "../../../components/ToggleButton.svelte";
-	import { get_variable_name } from "$lib/variables";
+	import SolutionFilter from '../../../components/SolutionFilter.svelte';
+	import type { ActivatedSolution, ComponentTotal, Dataset, Row } from '$lib/types';
+	import AllCheckbox from '../../../components/AllCheckbox.svelte';
+	import Radio from '../../../components/Radio.svelte';
+	import ToggleButton from '../../../components/ToggleButton.svelte';
+	import { get_variable_name } from '$lib/variables';
 
-	import { get_component_total } from "$lib/temple";
-	import { filter_and_aggregate_data, rename_field } from "$lib/utils";
-	import { tick } from "svelte";
-	import BarPlot from "../../../components/BarPlot.svelte";
+	import { get_component_total } from '$lib/temple';
+	import { filter_and_aggregate_data, rename_field } from '$lib/utils';
+	import { tick } from 'svelte';
+	import BarPlot from '../../../components/BarPlot.svelte';
+	import type { ChartConfiguration } from 'chart.js';
 
-	let carriers: string[] = [];
-	let nodes: string[] = [];
-	let years: number[] = [];
-	let locations: string[] = [];
-	let selected_locations: string[] = [];
-	let selected_solution: ActivatedSolution | null = null;
-	let selected_years: number[] = [];
-	let selected_normalisation: string = "not_normalized";
-	let selected_cost_carriers: string[] = [];
-	let selected_demand_carriers: string[] = [];
-	let solution_loading: boolean = false;
-	let fetching: boolean = false;
-	let transport_technologies: string[] = [];
-	let selected_transport_technologies: string[] = [];
-	let storage_technologies: string[] = [];
-	let selected_storage_technologies: string[] = [];
-	let conversion_technologies: string[] = [];
-	let selected_conversion_technologies: string[] = [];
+	let carriers: string[] = $state([]);
+	let nodes: string[] = $state([]);
+	let years: number[] = $state([]);
+	let locations: string[] = $state([]);
+	let selected_locations: string[] = $state([]);
+	let selected_solution: ActivatedSolution | null = $state(null);
+	let selected_years: number[] = $state([]);
+	let selected_normalisation: string = 'not_normalized';
+	let selected_cost_carriers: string[] = $state([]);
+	let selected_demand_carriers: string[] = $state([]);
+	let solution_loading: boolean = $state(false);
+	let fetching: boolean = $state(false);
+	let transport_technologies: string[] = $state([]);
+	let selected_transport_technologies: string[] = $state([]);
+	let storage_technologies: string[] = $state([]);
+	let selected_storage_technologies: string[] = $state([]);
+	let conversion_technologies: string[] = $state([]);
+	let selected_conversion_technologies: string[] = $state([]);
 	let grouped_data: Row[] | undefined;
 	let fetched_cost_carbon: ComponentTotal;
 	let fetched_cost_carrier: ComponentTotal;
-	let fetched_capex: ComponentTotal;
+	let fetched_capex: ComponentTotal | undefined = $state();
 	let fetched_opex: ComponentTotal;
 	let fetched_cost_shed_demand: ComponentTotal;
-	let cost_carriers: string[] = [];
-	let demand_carriers: string[] = [];
-	let combined_name = "Techology / Carrier";
-	let aggregation_options: string[] = [combined_name, "Location"];
-	let selected_aggregation: string = aggregation_options[1];
+	let cost_carriers: string[] = $state([]);
+	let demand_carriers: string[] = $state([]);
+	const combined_name = 'Techology / Carrier';
+	const aggregation_options: string[] = [combined_name, 'Location'];
+	let selected_aggregation: string = $state(aggregation_options[1]);
 
-	let show_costs = {
-		capex: { title: "Capex", show: true, subdivision: false },
-		opex: { title: "Opex", show: true, subdivision: false },
-		carrier: { title: "Carrier", show: true, subdivision: false },
-		shed_demand: { title: "Shed Demand", show: true, subdivision: false },
+	let show_costs = $state({
+		capex: { title: 'Capex', show: true, subdivision: false },
+		opex: { title: 'Opex', show: true, subdivision: false },
+		carrier: { title: 'Carrier', show: true, subdivision: false },
+		shed_demand: { title: 'Shed Demand', show: true, subdivision: false },
 		carbon_emission: {
-			title: "Carbon Emissions",
+			title: 'Carbon Emissions',
 			show: true,
-			subdivision: false,
-		},
-	};
+			subdivision: false
+		}
+	});
 
-	const capex_suffix = " (Capex)";
-	const opex_suffix = " (Opex)";
-	const capex_label = "Capex";
-	const opex_label = "Opex";
-	const carrier_label = "Carrier";
-	const shed_demand_label = "Shed Demand";
-	let plot_name = "plot";
+	const capex_suffix = ' (Capex)';
+	const opex_suffix = ' (Opex)';
+	const capex_label = 'Capex';
+	const opex_label = 'Opex';
+	const carrier_label = 'Carrier';
+	const shed_demand_label = 'Shed Demand';
+	let plot_name = $state('plot');
 
 	interface StringList {
 		[key: string]: string[];
 	}
 
 	// Define initial plot config
-	let config = {
-		type: "line",
-		data: { datasets: [] as any[] },
+	let filtered_data: any[] = $state([]);
+	let plot_config: ChartConfiguration = $derived({
+		type: 'line',
+		data: { datasets: filtered_data },
 		options: {
 			responsive: true,
 			scales: {
@@ -81,19 +78,19 @@
 					stacked: true,
 					title: {
 						display: true,
-						text: "Year",
-					},
+						text: 'Year'
+					}
 				},
 				y: {
 					stacked: true,
 					title: {
 						display: true,
-						text: "Costs [" + get_unit() + "]",
-					},
-				},
-			},
-		},
-	};
+						text: 'Costs [' + get_unit() + ']'
+					}
+				}
+			}
+		}
+	});
 
 	/**
 	 * This funciton fetches all necessary cost-data  given the currently selected solution.
@@ -107,54 +104,51 @@
 
 		fetched_capex = await get_component_total(
 			selected_solution!.solution_name,
-			get_variable_name("capex_yearly", selected_solution.version),
+			get_variable_name('capex_yearly', selected_solution.version),
 			selected_solution!.scenario_name,
 			selected_solution!.detail.system.reference_year,
-			selected_solution!.detail.system.interval_between_years,
+			selected_solution!.detail.system.interval_between_years
 		);
 
 		fetched_opex = await get_component_total(
 			selected_solution!.solution_name,
-			get_variable_name("opex_yearly", selected_solution.version),
+			get_variable_name('opex_yearly', selected_solution.version),
 			selected_solution!.scenario_name,
 			selected_solution!.detail.system.reference_year,
-			selected_solution!.detail.system.interval_between_years,
+			selected_solution!.detail.system.interval_between_years
 		);
 
 		fetched_cost_carbon = await get_component_total(
 			selected_solution!.solution_name,
-			get_variable_name(
-				"cost_carbon_emissions_total",
-				selected_solution.version,
-			),
+			get_variable_name('cost_carbon_emissions_total', selected_solution.version),
 			selected_solution!.scenario_name,
 			selected_solution!.detail.system.reference_year,
-			selected_solution!.detail.system.interval_between_years,
+			selected_solution!.detail.system.interval_between_years
 		);
 
 		fetched_cost_carrier = await get_component_total(
 			selected_solution!.solution_name,
-			get_variable_name("cost_carrier", selected_solution.version),
+			get_variable_name('cost_carrier', selected_solution.version),
 			selected_solution!.scenario_name,
 			selected_solution!.detail.system.reference_year,
-			selected_solution!.detail.system.interval_between_years,
+			selected_solution!.detail.system.interval_between_years
 		);
 
 		fetched_cost_shed_demand = await get_component_total(
 			selected_solution!.solution_name,
-			get_variable_name("cost_shed_demand", selected_solution.version),
+			get_variable_name('cost_shed_demand', selected_solution.version),
 			selected_solution!.scenario_name,
 			selected_solution!.detail.system.reference_year,
-			selected_solution!.detail.system.interval_between_years,
+			selected_solution!.detail.system.interval_between_years
 		);
 
 		// "Standardize" all series names
-		rename_field(fetched_cost_carrier.data!, "node", "location");
-		rename_field(fetched_cost_shed_demand.data!, "node", "location");
-		rename_field(fetched_opex.data!, "technology", combined_name);
-		rename_field(fetched_capex.data!, "technology", combined_name);
-		rename_field(fetched_cost_carrier.data!, "carrier", combined_name);
-		rename_field(fetched_cost_shed_demand.data!, "carrier", combined_name);
+		rename_field(fetched_cost_carrier.data!, 'node', 'location');
+		rename_field(fetched_cost_shed_demand.data!, 'node', 'location');
+		rename_field(fetched_opex.data!, 'technology', combined_name);
+		rename_field(fetched_capex.data!, 'technology', combined_name);
+		rename_field(fetched_cost_carrier.data!, 'carrier', combined_name);
+		rename_field(fetched_cost_shed_demand.data!, 'carrier', combined_name);
 
 		// Add Capex / Opex suffix
 		for (let i in fetched_capex.data!.data) {
@@ -168,7 +162,7 @@
 
 		// Set available locations
 		locations = Object.keys(selected_solution!.detail.edges).concat(
-			selected_solution!.detail.system.set_nodes,
+			selected_solution!.detail.system.set_nodes
 		);
 
 		fetching = false;
@@ -186,19 +180,16 @@
 		let filtered_cost_carrier = structuredClone(fetched_cost_carrier.data);
 
 		filtered_cost_carrier!.data = filtered_cost_carrier!.data.filter((i) =>
-			selected_cost_carriers.includes(i[combined_name]),
+			selected_cost_carriers.includes(i[combined_name])
 		);
 
-		let filtered_fetched_cost_shed_demand = structuredClone(
-			fetched_cost_shed_demand.data,
+		let filtered_fetched_cost_shed_demand = structuredClone(fetched_cost_shed_demand.data);
+
+		filtered_fetched_cost_shed_demand!.data = filtered_fetched_cost_shed_demand!.data.filter((i) =>
+			selected_demand_carriers.includes(i[combined_name])
 		);
 
-		filtered_fetched_cost_shed_demand!.data =
-			filtered_fetched_cost_shed_demand!.data.filter((i) =>
-				selected_demand_carriers.includes(i[combined_name]),
-			);
-
-		if (show_costs.capex.show) {
+		if (fetched_capex != undefined && show_costs.capex.show) {
 			// If the subdivision is active, no aggregation is necessary.
 			if (show_costs.capex.subdivision) {
 				grouped_data! = grouped_data!.concat(fetched_capex.data!.data);
@@ -208,10 +199,10 @@
 					{ location: selected_locations },
 					{ [combined_name]: all_selected_carriers_technologies },
 					[],
-					false,
+					false
 				);
 				for (let i in new_data) {
-					new_data[i].data["location"] = new_data[i].label;
+					new_data[i].data['location'] = new_data[i].label;
 					new_data[i].data[combined_name] = capex_label;
 					grouped_data.push(new_data[i].data);
 				}
@@ -227,13 +218,13 @@
 					fetched_opex.data!.data,
 					{ location: selected_locations },
 					{
-						[combined_name]: all_selected_carriers_technologies,
+						[combined_name]: all_selected_carriers_technologies
 					},
 					[],
-					false,
+					false
 				);
 				for (let i in new_data) {
-					new_data[i].data["location"] = new_data[i].label;
+					new_data[i].data['location'] = new_data[i].label;
 					new_data[i].data[combined_name] = opex_label;
 					grouped_data.push(new_data[i].data);
 				}
@@ -242,9 +233,7 @@
 
 		if (show_costs.carrier.show) {
 			if (show_costs.carrier.subdivision) {
-				grouped_data! = grouped_data!.concat(
-					filtered_cost_carrier!.data,
-				);
+				grouped_data! = grouped_data!.concat(filtered_cost_carrier!.data);
 			} else {
 				let comb = {};
 				let new_data = filter_and_aggregate_data(
@@ -252,11 +241,11 @@
 					{ location: selected_locations },
 					{ [combined_name]: all_selected_carriers_technologies },
 					[],
-					false,
+					false
 				);
 
 				for (let i in new_data) {
-					new_data[i].data["location"] = new_data[i].label;
+					new_data[i].data['location'] = new_data[i].label;
 					new_data[i].data[combined_name] = carrier_label;
 					grouped_data.push(new_data[i].data);
 				}
@@ -265,9 +254,7 @@
 
 		if (show_costs.shed_demand.show) {
 			if (show_costs.shed_demand.subdivision) {
-				grouped_data! = grouped_data!.concat(
-					filtered_fetched_cost_shed_demand!.data,
-				);
+				grouped_data! = grouped_data!.concat(filtered_fetched_cost_shed_demand!.data);
 			} else {
 				let comb = {};
 				let new_data = filter_and_aggregate_data(
@@ -275,10 +262,10 @@
 					{ location: selected_locations },
 					{ [combined_name]: all_selected_carriers_technologies },
 					[],
-					false,
+					false
 				);
 				for (let i in new_data) {
-					new_data[i].data["location"] = new_data[i].label;
+					new_data[i].data['location'] = new_data[i].label;
 					new_data[i].data[combined_name] = shed_demand_label;
 					grouped_data.push(new_data[i].data);
 				}
@@ -292,25 +279,27 @@
 			return;
 		}
 
+		// TODO: Find out when and why this actually was used
+		// Currently, this just removes labels once they are unselected, except if none are remaining.
 		let set_locations = new Set<string>();
 
 		// Filter undefined datasets
-		grouped_data! = grouped_data!.filter((e) => {
-			return e != undefined;
-		});
+		// grouped_data! = grouped_data!.filter((e) => {
+		// 	return e != undefined;
+		// });
 
-		for (const i of grouped_data!) {
-			set_locations.add(i["location"]);
-		}
+		// for (const i of grouped_data!) {
+		// 	set_locations.add(i['location']);
+		// }
 
 		// Update available locations
-		locations = Array.from(set_locations);
+		// locations = Array.from(set_locations);
 		fetching = false;
 	}
 
 	function get_unit() {
 		if (!fetched_capex || fetched_capex.unit === null) {
-			return "";
+			return '';
 		}
 
 		return fetched_capex.unit.data[0][0];
@@ -344,7 +333,7 @@
 		let all_selected_carriers_technologies = [
 			...suffixed_technologies,
 			...selected_cost_carriers,
-			...selected_demand_carriers,
+			...selected_demand_carriers
 		];
 
 		// Update the dataset aggregations and groupings
@@ -368,22 +357,18 @@
 
 		// Specify Aggregation. If the aggregation type is location, all technologies have to be aggregated and vice versa.
 		if (selected_aggregation == aggregation_options[0]) {
-			dataset_selector["location"] = locations;
-			datasets_aggregates[combined_name] =
-				all_selected_carriers_technologies;
+			dataset_selector['location'] = locations;
+			datasets_aggregates[combined_name] = all_selected_carriers_technologies;
 		} else {
-			dataset_selector[combined_name] =
-				all_selected_carriers_technologies;
-			datasets_aggregates["location"] = selected_locations;
+			dataset_selector[combined_name] = all_selected_carriers_technologies;
+			datasets_aggregates['location'] = selected_locations;
 		}
 
 		// Define excluded years
-		let excluded_years = years.filter(
-			(year) => !selected_years.includes(year),
-		);
+		let excluded_years = years.filter((year) => !selected_years.includes(year));
 
-		let normalized = selected_normalisation == "normalized";
-		let filtered_data: Dataset[] = [];
+		let normalized = selected_normalisation == 'normalized';
+		filtered_data = [];
 
 		// Get Plot-Data, as a base we now take the grouped data that are adapted to the subdivision selection.
 		if (grouped_data) {
@@ -392,30 +377,20 @@
 				dataset_selector,
 				datasets_aggregates,
 				excluded_years,
-				normalized,
+				normalized
 			);
 		}
 
 		// Get Total Carbon Cost Data
-		if (
-			fetched_cost_carbon.data!.data.length > 0 &&
-			show_costs.carbon_emission.show
-		) {
+		if (fetched_cost_carbon.data!.data.length > 0 && show_costs.carbon_emission.show) {
 			filtered_data.push({
-				label: "Total Carbon Costs",
+				label: 'Total Carbon Costs',
 				data: fetched_cost_carbon.data!.data[0],
-				type: "bar",
+				type: 'bar'
 			});
 		}
 
-		// Set the actual data in the plot config
-		tick().then(() => {
-			config.data = { datasets: filtered_data };
-
-			config.options.scales.y.title.text = "Costs [" + get_unit() + "]";
-
-			fetching = false;
-		});
+		fetching = false;
 	}
 
 	/**
@@ -433,30 +408,25 @@
 		// Options for Capex / Opex
 		let set_capex_opex_technologies = new Set<string>();
 
-		for (const row of fetched_capex.data!.data) {
-			set_capex_opex_technologies.add(
-				row[combined_name].replace(capex_suffix, ""),
-			);
+		if (fetched_capex != undefined) {
+			for (const row of fetched_capex.data!.data) {
+				set_capex_opex_technologies.add(row[combined_name].replace(capex_suffix, ''));
+			}
 		}
 		for (const row of fetched_opex.data!.data) {
-			set_capex_opex_technologies.add(
-				row[combined_name].replace(opex_suffix, ""),
-			);
+			set_capex_opex_technologies.add(row[combined_name].replace(opex_suffix, ''));
 		}
 
 		// Define available technologies of the different types
-		transport_technologies =
-			selected_solution!.detail.system.set_transport_technologies.filter(
-				(t) => set_capex_opex_technologies.has(t),
-			);
-		conversion_technologies =
-			selected_solution!.detail.system.set_conversion_technologies.filter(
-				(t) => set_capex_opex_technologies.has(t),
-			);
-		storage_technologies =
-			selected_solution!.detail.system.set_storage_technologies.filter(
-				(t) => set_capex_opex_technologies.has(t),
-			);
+		transport_technologies = selected_solution!.detail.system.set_transport_technologies.filter(
+			(t) => set_capex_opex_technologies.has(t)
+		);
+		conversion_technologies = selected_solution!.detail.system.set_conversion_technologies.filter(
+			(t) => set_capex_opex_technologies.has(t)
+		);
+		storage_technologies = selected_solution!.detail.system.set_storage_technologies.filter((t) =>
+			set_capex_opex_technologies.has(t)
+		);
 
 		// Reset selected technologies
 		selected_transport_technologies = transport_technologies;
@@ -483,8 +453,8 @@
 		for (const row of fetched_cost_shed_demand.data!.data) {
 			set_demand_carriers.add(row[combined_name]);
 		}
-		demand_carriers = selected_solution!.detail.carriers_demand.filter(
-			(i) => set_demand_carriers.has(i),
+		demand_carriers = selected_solution!.detail.carriers_demand.filter((i) =>
+			set_demand_carriers.has(i)
 		);
 
 		// Reset selected demand carriers
@@ -502,248 +472,228 @@
 		// Update the plot
 		update_plot_data();
 		fetching = false;
-		let solution_names = selected_solution!.solution_name.split(".");
+		let solution_names = selected_solution!.solution_name.split('.');
 
 		// Define the filename of the plot when downloading.
 		plot_name = [
 			solution_names[solution_names?.length - 1],
 			selected_solution?.scenario_name,
-			"costs",
-		].join("_");
+			'costs'
+		].join('_');
 	}
 </script>
 
 <h2>Costs</h2>
 
-<div class="row">
-	<div class="col position-relative" style="z-index: 1; position: relative;">
-		<div class="filters">
-			<div class="accordion" id="accordionExample">
-				<div class="accordion-item solution-selection">
+<div class="z-1 position-relative">
+	<div class="filters">
+		<div class="accordion" id="accordionExample">
+			<div class="accordion-item solution-selection">
+				<h2 class="accordion-header">
+					<button
+						class="accordion-button"
+						type="button"
+						data-bs-toggle="collapse"
+						data-bs-target="#collapseOne"
+						aria-expanded="true"
+						aria-controls="collapseOne"
+					>
+						Solution Selection
+					</button>
+				</h2>
+				<div
+					id="collapseOne"
+					class="accordion-collapse collapse show"
+					data-bs-parent="#accordionExample"
+				>
+					<div class="accordion-body">
+						<SolutionFilter
+							bind:carriers
+							bind:nodes
+							bind:years
+							bind:selected_solution
+							bind:loading={solution_loading}
+							solution_selected={solution_changed}
+							enabled={!solution_loading && !fetching}
+						/>
+					</div>
+				</div>
+			</div>
+			{#if !fetching && !solution_loading && fetched_capex && selected_solution != null}
+				<div class="accordion-item">
 					<h2 class="accordion-header">
 						<button
-							class="accordion-button"
+							class="accordion-button collapsed"
 							type="button"
 							data-bs-toggle="collapse"
-							data-bs-target="#collapseOne"
-							aria-expanded="true"
-							aria-controls="collapseOne"
+							data-bs-target="#collapseCost"
+							aria-expanded="false"
+							aria-controls="collapseCost"
 						>
-							Solution Selection
+							Cost selection
 						</button>
 					</h2>
 					<div
-						id="collapseOne"
-						class="accordion-collapse collapse show"
+						id="collapseCost"
+						class="accordion-collapse collapse"
 						data-bs-parent="#accordionExample"
 					>
 						<div class="accordion-body">
-							<SolutionFilter
-								on:solution_selected={solution_changed}
-								bind:carriers
-								bind:nodes
-								bind:years
-								bind:selected_solution
-								bind:loading={solution_loading}
-								enabled={!solution_loading && !fetching}
-							/>
+							{#each Object.keys(show_costs) as key}
+								<div class="d-flex">
+									<h4>{show_costs[key].title} :</h4>
+									<ToggleButton bind:value={show_costs[key].show} change={update_plot_data}
+									></ToggleButton>
+
+									{#if show_costs[key].show && key != 'carbon_emission'}
+										Subdivision
+										<ToggleButton bind:value={show_costs[key].subdivision} change={update_plot_data}
+										></ToggleButton>
+									{/if}
+								</div>
+							{/each}
 						</div>
 					</div>
 				</div>
-				{#if !fetching && !solution_loading && fetched_capex && selected_solution != null}
-					<div class="accordion-item">
-						<h2 class="accordion-header">
-							<button
-								class="accordion-button collapsed"
-								type="button"
-								data-bs-toggle="collapse"
-								data-bs-target="#collapseCost"
-								aria-expanded="false"
-								aria-controls="collapseCost"
-							>
-								Cost selection
-							</button>
-						</h2>
-						<div
-							id="collapseCost"
-							class="accordion-collapse collapse"
-							data-bs-parent="#accordionExample"
+				<div class="accordion-item">
+					<h2 class="accordion-header">
+						<button
+							class="accordion-button collapsed"
+							type="button"
+							data-bs-toggle="collapse"
+							data-bs-target="#collapseTwo"
+							aria-expanded="false"
+							aria-controls="collapseTwo"
 						>
-							<div class="accordion-body">
-								{#each Object.keys(show_costs) as key}
-									<div class="d-flex">
-										<h4>{show_costs[key].title} :</h4>
-										<ToggleButton
-											bind:value={show_costs[key].show}
-											on:change={update_plot_data}
-										></ToggleButton>
-
-										{#if show_costs[key].show && key != "carbon_emission"}
-											Subdivision
-											<ToggleButton
-												bind:value={show_costs[key]
-													.subdivision}
-												on:change={update_plot_data}
-											></ToggleButton>
-										{/if}
-									</div>
-								{/each}
-							</div>
+							Technology and Carrier Selection
+						</button>
+					</h2>
+					<div
+						id="collapseTwo"
+						class="accordion-collapse collapse"
+						data-bs-parent="#accordionExample"
+					>
+						<div class="accordion-body">
+							<h3>Technologies (for Capex/Opex)</h3>
+							{#if transport_technologies.length > 0}
+								<h4>Transport</h4>
+								<AllCheckbox
+									bind:selected_elements={selected_transport_technologies}
+									elements={transport_technologies}
+									selection_changed={() => {
+										update_plot_data();
+									}}
+								></AllCheckbox>
+							{/if}
+							{#if storage_technologies.length > 0}
+								<h4>Storage</h4>
+								<AllCheckbox
+									bind:selected_elements={selected_storage_technologies}
+									elements={storage_technologies}
+									selection_changed={() => {
+										update_plot_data();
+									}}
+								></AllCheckbox>
+							{/if}
+							{#if conversion_technologies.length > 0}
+								<h4>Conversion</h4>
+								<AllCheckbox
+									bind:selected_elements={selected_conversion_technologies}
+									elements={conversion_technologies}
+									selection_changed={() => {
+										update_plot_data();
+									}}
+								></AllCheckbox>
+							{/if}
+							{#if cost_carriers.length > 0}
+								<h3>Cost of Carrier</h3>
+								<AllCheckbox
+									bind:selected_elements={selected_cost_carriers}
+									elements={cost_carriers}
+									selection_changed={() => {
+										update_plot_data();
+									}}
+								></AllCheckbox>
+							{/if}
+							{#if demand_carriers.length > 0}
+								<h3>Shed Demand</h3>
+								<AllCheckbox
+									bind:selected_elements={selected_demand_carriers}
+									elements={demand_carriers}
+									selection_changed={() => {
+										update_plot_data();
+									}}
+								></AllCheckbox>
+							{/if}
 						</div>
 					</div>
-					<div class="accordion-item">
-						<h2 class="accordion-header">
-							<button
-								class="accordion-button collapsed"
-								type="button"
-								data-bs-toggle="collapse"
-								data-bs-target="#collapseTwo"
-								aria-expanded="false"
-								aria-controls="collapseTwo"
-							>
-								Technology and Carrier Selection
-							</button>
-						</h2>
-						<div
-							id="collapseTwo"
-							class="accordion-collapse collapse"
-							data-bs-parent="#accordionExample"
+				</div>
+				<div class="accordion-item">
+					<h2 class="accordion-header">
+						<button
+							class="accordion-button collapsed"
+							type="button"
+							data-bs-toggle="collapse"
+							data-bs-target="#collapseThree"
+							aria-expanded="false"
+							aria-controls="collapseThree"
 						>
+							Data Selection
+						</button>
+					</h2>
+					<div
+						id="collapseThree"
+						class="accordion-collapse collapse"
+						data-bs-parent="#accordionExample"
+					>
+						<div class="accordion-body">
 							<div class="accordion-body">
-								<h3>Technologies (for Capex/Opex)</h3>
-								{#if transport_technologies.length > 0}
-									<h4>Transport</h4>
-									<AllCheckbox
-										bind:selected_elements={selected_transport_technologies}
-										bind:elements={transport_technologies}
-										on:selection-changed={() => {
-											update_plot_data();
-										}}
-									></AllCheckbox>
-								{/if}
-								{#if storage_technologies.length > 0}
-									<h4>Storage</h4>
-									<AllCheckbox
-										bind:selected_elements={selected_storage_technologies}
-										bind:elements={storage_technologies}
-										on:selection-changed={() => {
-											update_plot_data();
-										}}
-									></AllCheckbox>
-								{/if}
-								{#if conversion_technologies.length > 0}
-									<h4>Conversion</h4>
-									<AllCheckbox
-										bind:selected_elements={selected_conversion_technologies}
-										bind:elements={conversion_technologies}
-										on:selection-changed={() => {
-											update_plot_data();
-										}}
-									></AllCheckbox>
-								{/if}
-								{#if cost_carriers.length > 0}
-									<h3>Cost of Carrier</h3>
-									<AllCheckbox
-										bind:selected_elements={selected_cost_carriers}
-										bind:elements={cost_carriers}
-										on:selection-changed={() => {
-											update_plot_data();
-										}}
-									></AllCheckbox>
-								{/if}
-								{#if demand_carriers.length > 0}
-									<h3>Shed Demand</h3>
-									<AllCheckbox
-										bind:selected_elements={selected_demand_carriers}
-										bind:elements={demand_carriers}
-										on:selection-changed={() => {
-											update_plot_data();
-										}}
-									></AllCheckbox>
-								{/if}
-							</div>
-						</div>
-					</div>
-					<div class="accordion-item">
-						<h2 class="accordion-header">
-							<button
-								class="accordion-button collapsed"
-								type="button"
-								data-bs-toggle="collapse"
-								data-bs-target="#collapseThree"
-								aria-expanded="false"
-								aria-controls="collapseThree"
-							>
-								Data Selection
-							</button>
-						</h2>
-						<div
-							id="collapseThree"
-							class="accordion-collapse collapse"
-							data-bs-parent="#accordionExample"
-						>
-							<div class="accordion-body">
-								<div class="accordion-body">
-									<div class="row">
-										<div class="col-6">
-											<h3>Aggregation</h3>
-											<Radio
-												bind:options={aggregation_options}
-												bind:selected_option={selected_aggregation}
-												on:selection-changed={(e) => {
-													update_plot_data();
-												}}
-											></Radio>
-										</div>
+								<div class="row">
+									<div class="col-6">
+										<h3>Aggregation</h3>
+										<Radio
+											options={aggregation_options}
+											bind:selected_option={selected_aggregation}
+											selection_changed={(_) => update_plot_data()}
+										></Radio>
 									</div>
-									{#if selected_aggregation == aggregation_options[1]}
-										<h3>Location</h3>
-										<AllCheckbox
-											bind:selected_elements={selected_locations}
-											elements={locations}
-											on:selection-changed={(e) => {
-												update_plot_data();
-											}}
-										></AllCheckbox>
-									{/if}
-									{#if selected_years}
-										<h3>Year</h3>
-										<AllCheckbox
-											bind:selected_elements={selected_years}
-											elements={years}
-											on:selection-changed={(e) => {
-												update_plot_data();
-											}}
-										></AllCheckbox>
-									{/if}
 								</div>
+								{#if selected_aggregation == aggregation_options[1]}
+									<h3>Location</h3>
+									<AllCheckbox
+										bind:selected_elements={selected_locations}
+										elements={locations}
+										selection_changed={(_) => update_plot_data()}
+									></AllCheckbox>
+								{/if}
+								{#if selected_years}
+									<h3>Year</h3>
+									<AllCheckbox
+										bind:selected_elements={selected_years}
+										elements={years}
+										selection_changed={(_) => update_plot_data()}
+									></AllCheckbox>
+								{/if}
 							</div>
 						</div>
 					</div>
-				{/if}
-			</div>
+				</div>
+			{/if}
 		</div>
 	</div>
 </div>
-<div class="row">
-	<div class="col mt-4">
-		{#if solution_loading || fetching}
-			<div class="text-center">
-				<div class="spinner-border center" role="status">
-					<span class="visually-hidden">Loading...</span>
-				</div>
+<div class="mt-4">
+	{#if solution_loading || fetching}
+		<div class="text-center">
+			<div class="spinner-border center" role="status">
+				<span class="visually-hidden">Loading...</span>
 			</div>
-		{:else if selected_solution != null}
-			{#if config.data.datasets.length == 0 || config.data.datasets[0].data.length == 0}
-				<div class="text-center">No data with this selection.</div>
-			{:else}
-				<BarPlot
-					bind:config
-					bind:year_offset={selected_solution.detail.system
-						.reference_year}
-					bind:plot_name
-				></BarPlot>
-			{/if}
+		</div>
+	{:else if selected_solution != null}
+		{#if filtered_data.length == 0 || filtered_data[0].data.length == 0}
+			<div class="text-center">No data with this selection.</div>
+		{:else}
+			<BarPlot config={plot_config} {plot_name}></BarPlot>
 		{/if}
-	</div>
+	{/if}
 </div>
